LatinIME: Move keyboard height pref out of debug
Move the preference out of debug into advanced options.

Adapted from https://github.com/dslul/openboard

Test:
Compiles,
The preferences are visible in "Advanced" preferences,
The height of the keyboard follows the preference value,
The height is reset to default value when "Enable keyboard resizing" is disabled.

Also expose the strings for translation.

Change-Id: I68aebcc4d1c26b6c4b648cac333285c496990b27